# api/swagger.yaml
basePath: /
definitions:
  db.OrderRecipe:
    properties:
      created_at:
        type: string
      employee_id:
        type: string
      id:
        type: string
      order_id:
        type: string
      recipe_id:
        type: string
      removed_at:
        type: string
    type: object
...
  httpx.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  httpx.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/httpx.APIError'
    type: object
...
  orders.AttachRecipeRequest:
    properties:
      employeeId:
        type: string
      recipeId:
        type: string
    type: object
...
  recipes.BatchRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
  recipes.ComponentInput:
    properties:
      brand:
        example: DS
        type: string
      flavour:
        example: apple
        type: string
      percent:
        example: 20
        type: integer
    type: object
  recipes.ComponentView:
    properties:
      brand:
        type: string
      flavour:
        type: string
      percent:
        type: integer
    type: object
  recipes.CreateRecipeRequest:
    properties:
      components:
        items:
          $ref: '#/definitions/recipes.ComponentInput'
        type: array
      isSecret:
        type: boolean
      name:
        example: Apple Mix
        type: string
      strength:
        example: 6
        type: integer
    type: object
  recipes.RecipeView:
    properties:
      components:
        items:
          $ref: '#/definitions/recipes.ComponentView'
        type: array
      id:
        type: string
      isSecret:
        type: boolean
      name:
        type: string
      strength:
        type: integer
    type: object
  users.AddFavouriteRequest:
    properties:
      orderRecipeId:
        type: string
    type: object
  users.ComponentView:
    properties:
      brand:
        type: string
      flavour:
        type: string
      percent:
        type: integer
    type: object
  users.FavouriteView:
    properties:
      authorFullName:
        type: string
      authorShortName:
        type: string
      components:
        items:
          $ref: '#/definitions/users.ComponentView'
        type: array
      isSecret:
        type: boolean
      likedAt:
        type: string
      myReview:
        type: string
      myScore:
        type: integer
      orderRecipeId:
        type: string
      recipeId:
        type: string
      recipeName:
        type: string
      restaurantId:
        type: string
      restaurantName:
        type: string
      strength:
        type: integer
    type: object
  users.RegisterRequest:
    properties:
      gender:
        example: male
        type: string
      phoneNumber:
        example: "+79991234567"
        type: string
    type: object
  users.UserResponse:
    properties:
      createdAt:
        type: string
      gender:
        type: string
      id:
        type: string
      phoneNumber:
        type: string
    type: object
info:
  contact: {}
  description: Backend for the mixMaster restaurant ordering / recipes app.
  title: mixMaster API
  version: 0.1.0
paths:
...
  /orders/{id}/recipes:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Recipe + author
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/orders.AttachRecipeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/db.OrderRecipe'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
        "409":
          description: order is closed
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Attach a recipe to an order
      tags:
      - orders
  /orders/{id}/recipes/{orderRecipeId}:
    delete:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: OrderRecipe ID
        in: path
        name: orderRecipeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Remove/replace a recipe on an order (soft, removed_at)
      tags:
      - orders
...
  /recipes:
    post:
      consumes:
      - application/json
      parameters:
      - description: Recipe + components
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/recipes.CreateRecipeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/recipes.RecipeView'
              type: object
        "400":
          description: validation (e.g. percent != 100)
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Create a recipe with components (must sum to 100%)
      tags:
      - recipes
  /recipes/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Recipe ids
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/recipes.BatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/recipes.RecipeView'
                  type: array
              type: object
      summary: Get recipes by ids (batch), components included
      tags:
      - recipes
...
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: Guest data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/users.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Register a guest
      tags:
      - users
  /users/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/users.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.Envelope'
      summary: Get guest data
      tags:
      - users
  /users/{id}/favourites:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/users.FavouriteView'
                  type: array
              type: object
      summary: 'List guest favourites (enriched: restaurant, components, author)'
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: OrderRecipe to like
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.AddFavouriteRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Add a recipe to favourites
      tags:
      - users
  /users/{id}/favourites/{orderRecipeId}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: OrderRecipe ID
        in: path
        name: orderRecipeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Remove a recipe from favourites
      tags:
      - users
swagger: "2.0"